+++ src/server/createSSR.js
@@ -14,10 +14,11 @@ import {Map as iMap} from 'immutable';
 // https://github.com/systemjs/systemjs/issues/953
 
 function renderApp(res, store, assets, renderProps) {
-  const location = renderProps ? renderProps.location : '/';
+  const location = renderProps && renderProps.location && renderProps.location.pathname || '/';
   // Needed so some components can render based on location
   store.dispatch(push(location));
   const htmlStream = renderToStaticMarkup(<Html
+    title="Action | Parabol Inc"
     store={store}
     assets={assets}
     renderProps={renderProps}
@@ -35,7 +36,7 @@ export default async function createSSR(req, res) {
     const assets = require('../../build/assets.json');
     const readFile = promisify(fs.readFile);
     assets.manifest.text = await readFile(join(__dirname, '..', '..', 'build', basename(assets.manifest.js)), 'utf-8');
-    const routes = makeRoutes(store);
+      const routes = makeRoutes(store);
     match({routes, location: req.url}, (error, redirectLocation, renderProps) => {
       if (error) {
         res.status(500).send(error.message);