+++ src/server/graphql/rootQuery.js
@@ -2,7 +2,7 @@ import {GraphQLObjectType} from 'graphql';
 import cachedUser from './models/CachedUser/cachedUserQuery';
 import meeting from './models/Meeting/meetingQuery'
 
-const rootFields = {cachedUser, meeting};
+const rootFields = Object.assign(cachedUser);
 
 export default new GraphQLObjectType({
   name: 'RootQuery',