+++ src/universal/routes/landing.js
@@ -1,11 +1,12 @@
 // import Home from 'universal/modules/landing/components/Home/Home';
 import LandingContainer from 'universal/modules/landing/containers/Landing/LandingContainer';
+import Landing from 'universal/modules/landing/components/Landing/Landing';
 
 export default {
   path: '/',
   component: LandingContainer,
-  // indexRoute: {
-  //   component: Home
-  // },
-  childRoutes: []
+  indexRoute: {
+    component: Landing
+  }
+  // childRoutes: []
 };