+++ webpack/webpack.config.dev.js
@@ -6,11 +6,7 @@ const root = process.cwd();
 const clientInclude = [path.join(root, 'src', 'client'), path.join(root, 'src', 'universal')];
 const globalCSS = path.join(root, 'src', 'universal', 'styles', 'global');
 
-const prefetches = [
-  'react-dnd/lib/index.js',
-  'joi/lib/index.js',
-  'universal/modules/kanban/containers/Kanban/KanbanContainer.js'
-];
+const prefetches = [];
 
 const prefetchPlugins = prefetches.map(specifier => new webpack.PrefetchPlugin(specifier));