+++ CHANGELOG.md
@@ -11,7 +11,10 @@ This CHANGELOG follows conventions [outlined here](http://keepachangelog.com/).
 - Removed sessions
 - Removed Sails
 - Switched from Falcor to GraphQL
-- 
+- Upgrade to Babel 6
+- Upgrade to react-router
+- Upgrade to react-router-redux
+
 ## v0.0.1
 ### Added
 - Initial fork from